+++ web/app/config/routing.yml
@@ -18,92 +18,3 @@ logout:
 
 error403:
     pattern: /403
-    
-profile_page:
-    path: /profile
-    defaults: { _controller: AdminBundle:Default:profile }
-
-emp_page_w_params:
-    path: /emp/{id}
-    defaults: { _controller: AdminBundle:Default:employeeProfile }
-
-emp_page:
-    path: /emp/
-    defaults: { _controller: AdminBundle:Default:employeeProfile }
-
-view_request:
-    path: /requests
-    defaults: { _controller: AdminBundle:Default:request }
-
-manage_employee:
-    path: /manage
-    defaults: { _controller: AdminBundle:Default:manage }
-
-add_employee:
-    path: /addemp
-    defaults: { _controller: AdminBundle:Default:addEmployee }
-
-add_position:
-    path: /addpos/
-    defaults: { _controller: AdminBundle:Default:addPosition }
-
-add_position_w_params:
-    path: /addpos/{id}
-    defaults: { _controller: AdminBundle:Default:addPosition }
-
-add_department:
-    path: /adddept/
-    defaults: { _controller: AdminBundle:Default:addDepartment }
-
-add_department_w_params:
-    path: /adddept/{id}
-    defaults: { _controller: AdminBundle:Default:addDepartment }
-
-notif_page:
-    path: /notif
-    defaults: { _controller: AdminBundle:Default:notif }
-
-emp_delete:
-    path: /delete/
-    defaults: { _controller: AdminBundle:Default:empDelete }
-
-emp_delete_w_params:
-    path: /delete/{id}
-    defaults: { _controller: AdminBundle:Default:empDelete }   
-
-update_profile:
-    path: /update
-    defaults: { _controller: AdminBundle:Default:profileUpdate } 
-
-request_access_accepted:
-    path: /requestaccepted
-    defaults: { _controller: AdminBundle:Default:acceptRequest }
-
-change_pass:
-    path: /changepass
-    defaults: {_controller: AdminBundle:Default:changePassword }
-
-check_timein:
-    path: /check_timein
-    defaults: {_controller: AdminBundle:Default:checkTimeIn }
-
-send_email:
-    path: /sendemail/
-    defaults: {_controller: AdminBundle:Default:sendEmail }
-
-export:
-    path: /export
-    defaults: {_controller: AdminBundle:Default:export }
-
-request_access:
-    path: /requestaccess
-    defaults: {_controller: AdminBundle:Default:requestAccess }
-
-
-add_event_accepted_request:
-    path: /addevent/acceptrequest
-    defaults: { _controller:AdminBundle:Default:addRequestCalendar }
-
-add_event_timed_in_today:
-    path: /addevent/timedin
-    defaults: { _controller:AdminBundle:Default:timedInTodayCalendar }
