+++ web/src/AdminBundle/Resources/views/Default/index.html.twig
@@ -82,7 +82,7 @@
                                     </div>
                                 </div>
                             </div>
-                            <a class="btn-floating btn-large waves-effect waves-light red" id="add-request-form" style = "float:right; margin-right: 10px; display: none;"><i class="material-icons">add</i></a>
+                            <a class="btn-floating btn-large waves-effect waves-light red" id="add-request-form" style = "float:right; margin-right: 10px;"><i class="material-icons">add</i></a>
                         </div>
                     </div>
