+++ web/src/AdminBundle/Resources/views/Employee/manage.html.twig
@@ -12,7 +12,7 @@
                         <div class="responsive-nav"><a class="btn-flat"><i class="material-icons res-menu">reorder</i></a></div>
                         <div class="top-nav">
                             <h3>Employee List</h3>
-                            {{ include ('AdminBundle:Default:manageemployees.html.twig') }}
+                            {{ include ('AdminBundle:Employee:manageemployees.html.twig') }}
                         </div>
                     </div>
                     <div class="col-md-12 content manageemployees-content" style="display: block;">
