+++ web/src/AdminBundle/Resources/views/Employee/managetime.html.twig
@@ -22,7 +22,7 @@
                             <div class="responsive-nav"><a class="btn-flat"><i class="material-icons res-menu">reorder</i></a></div>
                             <div class="top-nav">
                                 <h3>Employee Time Logs</h3>
-                                {{ include ('AdminBundle:Default:manageemployees.html.twig') }}
+                                {{ include ('AdminBundle:Employee:manageemployees.html.twig') }}
                             </div>
                         </div>
                         <div class="col-md-12 content managetime-content">
