--- web/src/AdminBundle/Resources/views/EventManager/Manage.html.twig
@@ -1,116 +1,0 @@
-{% extends 'AdminBundle::base.html.twig' %}
-{% block title %}
-    <title>Propelrr Log-in System| Manage Events</title>
-{% endblock %}
-
-
-{% block body %}
-    {{ notif.add_notif }}
-    {{ notif.notif_loader }}
-    <section id="main-container">
-        <section id="main-wrapper">
-            <div class="container-fluid">
-                <div class="row">
-                    <div class="page-content">
-                        <div class="col-md-12 light-blue darken-1 heading z-depth-1">
-                            <div class="responsive-nav"><a class="btn-flat"><i class="material-icons res-menu">reorder</i></a></div>
-                            <div class="top-nav">
-                                <h3>Events List</h3>
-                                <a class="waves-effect waves-light btn amber btn-add" id = "btn-add-event" style="color: white;">
-                                    Create Event</a>
-                            </div>
-                        </div>
-                    </div>
-                </div>
-            </div>
-        </section>
-    </section>
-
-    <!-----------------------DAILY AGENDA MODAL -------------------------->
-    <div id="add_event_modal" class="modal modal-fixed-footer">
-
-        <div class="modal-content">
-            <h4 class="agenda-date">Create Event</h4>
-            <br>
-            <div class = "event-form-cont">
-                <div class = "row">
-                    <div class = "col-md-4">
-                        <span class="event-lbl">Date</span>
-                        <input type = "text" name="event-date" readonly placeholder="Event Date" class="event-date datepicker" id="event_date">
-                    </div>
-                    <div class = "col-md-8">
-                        <span class="event-lbl">Name </span>
-                        <input type = "text" name="event-name" placeholder="Event Name" class="event-name" id="event_name" required>
-                    </div>
-                </div>
-            </div>
-
-            <div class = "row">
-                <div class = "col-md-12">
-                    <textarea id="event_desc" class="materialize-textarea" placeholder="Event Description" required autocapitalize></textarea>
-                </div>
-            </div>
-
-            <!--<div class = "row">
-                <div class = "col-md-4">
-                    <span class="event-lbl">Event Type</span>
-                </div>
-                <div class = "col-md-3">
-                    <p>
-                        <input type="radio" class="filled-in" id="holiday" value="holiday" name = "event-type" />
-                        <label for="holiday" class="black-text">Holiday</label>
-                    </p>
-                </div>
-                <div class = "col-md-3">
-                    <p>
-                        <input type="radio" class="filled-in" id="custom" value="custom" name = "event-type"/>
-                        <label for="custom" class="black-text">Custom Event</label>
-                    </p>
-                </div>
-            </div>-->
-
-            <div class = "row">
-                <div class = "event-type-group">
-                    <div class="btn-group btn-group-justified" role="group" aria-label="...">
-                        <div class="btn-group" role="group">
-                            <button type="button" class="btn btn-default">Left</button>
-                        </div>
-                        <div class="btn-group" role="group">
-                            <button type="button" class="btn btn-default">Right</button>
-                        </div>
-                    </div>
-                </div>
-            </div>
-
-            <br>
-        </div>
-
-        <div class="modal-footer">
-            <a href="#!" class="waves-effect waves-light grey custombtn btn-flat close-modal-agenda" id = "btn-close">Close</a>
-            <a class="waves-effect waves-light btn amber btn-create" style="margin-right:5px;">Create</a>
-        </div>
-    </div>
-    <!-----------------------END OF MODAL -------------------------->
-{% endblock %}
-
-{% block customjs %}
-    <script type="text/javascript">
-        $(document).ready(function(){
-            $("#btn-add-event").click(function() {
-                $('#add_event_modal').openModal();
-            });
-
-            $(".btn-create").click(function() {
-                create();
-            });
-
-            $("#btn-close").click(function() {
-                $('#add_event_modal').closeModal();
-            });
-        });
-
-        function create() {
-
-        }
-    </script>
-{% endblock %}
