+++ web/web/css/event-style.css
@@ -1,0 +1,28 @@
+.event-type {
+    background: none;
+    border: none;
+    border-bottom: 1px solid #e0e0e0;
+}
+
+.option-type {
+    padding: 2px 5px;
+    background: none;
+}
+
+#view_event_desc {
+    word-wrap: break-word;
+}
+
+.desc-cont {
+    border-bottom: 1px solid #9e9e9e;
+    margin: 0 0 15px 0;
+    padding-bottom: 10px;
+}
+
+.event-desc {
+    display:none;
+}
+
+.event-btn {
+    margin-right: 10px; !important;
+}
