+++ web/web/prols2/css/style.css
@@ -3706,10 +3706,9 @@ label {
 
 label {
     font-size: 14px;
-    text-transform: uppercase;
-    font-weight: 500;
-    color: white;
-    padding: 7px;
+    text-transform: none;
+    color: #757575;
+    padding: 0px;
     border-radius: 5%
 }
